[10.0] hr_expense_usability improvements (#37)
* [FIX] hr_expense_usability: set default move number

Otherwise it's not possible to post the account move.

* [FIX] hr_expense_usability: prevent deleting draft move

This prevents deleting a draft account move that is linked
to an expense. Without this the expense could stay in posted
while the related move has been deleted, which is inconsitent
and impossible to recover for the user.

* [IMP] hr_expense_usability: preserve product/quantity on move line

When possible, preserve product and quantity on move line. This is
important when generating analytic lines to reinvoice expenses

* add contributor

* [FIX] hr_expense_usability: readonly account move on expense sheet

* [IMP] hr_expense_usablility: preserve employee name in move line name

This is important when reinvoicing expenses, so the correct label
is visible on the sale order.